requirements.txt
Django==1.9.7
django-annoying==0.9.0
MySQL-python==1.2.5
